+++ tests/Application.UnitTests/CommandTests/Stubs/StubStringLocalizer.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.Localization;
+
+namespace ShopSharp.Users.Domain.CommandTests.Stubs;
+
+internal class StubStringLocalizer<T> : IStringLocalizer<T>
+{
+    public LocalizedString this[string name] => new(name, name);
+
+    public LocalizedString this[string name, params object[] arguments] => new(name, string.Format(name, arguments));
+
+    public IEnumerable<LocalizedString> GetAllStrings(bool includeParentCultures)
+    {
+        throw new NotImplementedException();
+    }
+}