+++ tests/Domain.UnitTests/Usings.cs
@@ -1,4 +1,0 @@
-global using AutoFixture.Xunit2;
-global using CSharpFunctionalExtensions.FluentAssertions;
-global using FluentAssertions;
-global using Xunit;